requirements.txt
# Credit Risk Modeling - Dependencies

# Core Data Science
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.7.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Workflow Orchestration (Bonus)
metaflow>=2.9.0

# AWS (Production)
boto3>=1.26.0
# psycopg2-binary>=2.9.0  # For Redshift (uncomment in prod)

# Utilities
jupyter>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0

# Optional: Additional ML libraries
# lightgbm>=3.3.0
# shap>=0.41.0
